+++ docs/conf.py
@@ -25,7 +25,7 @@ copyright = '2017-2020, Sage Bionetworks'
 author = 'Thomas Yu'
 
 # The short X.Y version
-version = '2.1'
+version = '2.2'
 # The full version, including alpha/beta/rc tags
 release = version